requirements.txt
numpy
PyQt6
pyqtgraph
opencv-python
pyzmq
msgpack